// test/integration/test/mongo/where-filter/_fixture/generated/contract.d.ts
export type StorageHash =
  StorageHashBase<'bf3970799693599d69b578642e922405a9433bf6e3b1723856ac304ae73c2cf7'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'251b3ce23f6c9f561892e7c1af9d2cc941a13d64ba1aa7226b90036b09568cc3'>;

export type CodecTypes = MongoCodecTypes;

export type FieldOutputTypes = {
  readonly __unbound__: {
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly name: CodecTypes['mongo/string@1']['output'] | null;
      readonly tags: ReadonlyArray<CodecTypes['mongo/string@1']['output']>;
    };
  };
};
export type FieldInputTypes = {
  readonly __unbound__: {
    readonly User: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly name: CodecTypes['mongo/string@1']['input'] | null;
      readonly tags: ReadonlyArray<CodecTypes['mongo/string@1']['input']>;
    };
  };
};

export namespace Models {
  export type unbound_User = {
    _id: CodecTypes['mongo/objectId@1']['output'];
    name: CodecTypes['mongo/string@1']['output'] | null;
    tags: ReadonlyArray<CodecTypes['mongo/string@1']['output']>;
    readonly [RelationKeys]?: never;
  };
}

export declare const models: {
  __unbound__: {
    User: Models.unbound_User;
  };
};

export type TypeMaps = MongoTypeMaps<CodecTypes, FieldOutputTypes, FieldInputTypes>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly __unbound__: {
        readonly id: '__unbound__';
        readonly kind: 'mongo-database';
        readonly entries: {
          readonly collection: {
            readonly users: {
              readonly kind: 'mongo-collection';
              readonly validator: {
                readonly jsonSchema: {
                  readonly additionalProperties: false;
                  readonly bsonType: 'object';
                  readonly properties: {
                    readonly _id: { readonly bsonType: 'objectId' };
                    readonly name: { readonly bsonType: readonly ['null', 'string'] };
                    readonly tags: {
                      readonly bsonType: 'array';
                      readonly items: { readonly bsonType: 'string' };
                    };
                  };
                  readonly required: readonly ['_id', 'tags'];
                };
                readonly kind: 'mongo-validator';
                readonly validationAction: 'error';
                readonly validationLevel: 'strict';
              };
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'mongo';
  readonly targetFamily: 'mongo';
  readonly roots: {
    readonly users: { readonly namespace: '__unbound__' & NamespaceId; readonly model: 'User' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly __unbound__: {
        readonly models: {
          readonly User: {
            readonly fields: {
              readonly _id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'mongo/objectId@1' };
              };
              readonly name: {
                readonly nullable: true;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'mongo/string@1' };
              };
              readonly tags: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'mongo/string@1' };
                readonly many: true;
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: { readonly collection: 'users' };
          };
        };
      };
    };
  };
  readonly capabilities: {};
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = MongoContractWithTypeMaps<ContractBase, TypeMaps>;